rename api folder to collections.
in reference to mongo collections
