+++ RxOn/imports/collections/tasks.js
@@ -1,8 +1,4 @@
-
-
 import { Mongo } from 'meteor/mongo';
-
- 
 
 export const Tasks = new Mongo.Collection('tasks');
